Add headers and trailers to TapEvents (#33)
### Motivation

In order to include headers and trailers in `TapEvent`s, the API must be updated
to allow a new field to be set in the proxy. This field should be set by the
proxies and read by the control plane.

### Solution

A new message `http_types.Headers` has been added to `http_types.proto`. This
message allows repeated fields of `name` and `value` pairs where `name` can
non-unique. Because header values can be encoded in various ways, `value` is
`bytes` while `name` is `string`

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>