rust: add Tonic version of generated Rust code
Tonic (https://crates.io/crates/tonic) is, essentially, the Tokio 0.2
world's version of `tower-grpc`. Although the two libraries are similar
and share a common geneology, the code generated by `tonic-build` is not
compatible with the code generated by `tower-grpc-build`. In order to
implement the mock destination controller for linkerd/linkerd2#4158
using the Tokio 0.2 stack, we need a Tonic version of the generated Rust
bindings.

This branch adds a new `linkerd2-proxy-api-tonic` crate that contains
Tonic bindings generated from the proxy-api protos. Rather than putting
this code in the existing `linkerd2-proxy-api` crate under a feature
flag, I opted to create a separate crate. This is because `tonic-build`
and `tower-grpc-build` will generate code that depends on different
versions of libraries like `tower`, `http`, and `prost`. Although it
would be possible to rename those dependencies so we can depend on both
their Tokio 0.1 and Tokio 0.2 versions, it seems significantly more
difficult to convince the code generators to generate code that
references those dependencies by names like `http_02` and so on.
Therefore, I created a separate crate for the Tonic version.

This may also be useful when we start updating the proxy itself to use
the Tokio 0.2 stack.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>