rust: Add a SocketAddr converter (#46)
There is no `TryFrom` implementation from `net.TcpAddress` to
`SocketAddr`,  though we've manually implemented it in the
`linkerd2-api-resovle` crate.

This change provides a generalized implementation. It also modifies
existing IP converters to be on the address type directly and not a
reference (since IP addrs are generally considered copyable).