destination: Add opaque_protocol and fully_qualified_name to profile response (#44)
The profile API can be used to bootstrap dynamic protocol detection.
This change adds an `opaque_protocol` flag to the `DestinationProfile`,
which indicates that the proxy should skip protocol detection. HTTP
routes will not be used, but `dst_overrides` will.

`fully_qualified_name` provides returns the name of the resolved service.
This is useful when resolutions are performed on IP addresses.