dst: Indicate endpoint support for opaque transport (#53)
The proxy supports reading a connection header for opaque traffic. The
inbound proxy reads this header from connections that target the proxy's
inbound port; but outbound proxies have no means to know whether a given
traffic target supports this header and on what port it expects the
connection header. (While it's almost always 4143, it's theoretically
configurable.)

This change adds an optional field to an endpoint's `ProtocolHint` that
should be set when the target endpoint supports transporting opaque
traffic via the connection header.