Remove Cargo.lock (#74)
We store Cargo.lock so that local tests are consistent with those in CI,
but this leads to false positive auditing alerts that cannot be fixed by
dependabot.

This change removes Cargo.lock from version control to avoid these
auditing issues. This is acceptable, as this workspace only provides
library code. Latest dependency versions will always be used in CI.