Move rust sources into the repo's root
Currently, the core project is a cargo workspace with a single package
in the `rs/` directory containing the package's Cargo.toml and sources.
This prevents the crate from being published to crates.io, since the
`protos` dir is not part of the package directory.

To fix this, this change moves the contents of the `rs/` directory to
the root of repo, replacing the cargo worksapce with a single cargo
package.