Add a devcontainer config
Add a devcontainer config, reusing the container image from the main
linkerd2 repo.

This change replaces the `rust-toolchain` file, which forces use of a
single rust version when interacting with the repo, with a documented
minimum supported `rust-version` in `Cargo.toml`. This allows the dev
image's rust version to work properly when running `cargo`, etc.

Signed-off-by: Oliver Gould <ver@buoyant.io>