rust: Remove build dependency on `protoc`
The protobuf project does not publish `protoc` binaries for 32-bit ARM
hosts, which complicates our compilation tooling for ARM. This problem
can be avoided completely by committing the generated interface
bindings so that crates that depend on `linkerd2-proxy-api` don't need
any build dependencies.

This change removes the `client`, `server`, and `transport` features.
Client and server bindings are now always generated for gRPC services.
The `transport` feature is removed, since it cannot be optional if the
generated code is stored in the repo. We don't want to require tonic's
transport feature because it pulls in many dependencies that we don't
really need.

Signed-off-by: Oliver Gould <ver@buoyant.io>