rust: Add a Cargo.lock file
Because this crate is a library, we did not include the `Cargo.lock`
file in version control. However, this causes unexpected CI breakages,
since each CI run may pull new dependencies that are subtly
incompatible.

For example, changes to generated code formatting can cause CI to fail.

This change commits the `Cargo.lock` file to version control so we can
have a predictable integration environment.

Signed-off-by: Oliver Gould <ver@buoyant.io>