Add a `Metadata` protobuf module (#123)
As we introduce HTTP routes (for inbound policy), we need the control
plane to describe resources references for metrics, and other
diagonstics in the proxy. Currently, the control plane uses a free-form
map of strings, but the proxy only references a specific set of keys,
which is not clearly documented via the gRPC API.

This change replaces the `Authz` type's `labels` map with a `Metadata`
type. Controllers should set both fields for at least one stable
version.

Signed-off-by: Oliver Gould <ver@buoyant.io>