rust: actually include `outbound` module
PR #165 adds a proposal for the new `outbound` API for discovering
client-side HTTPRoute policy. However, this branch cannot currently be
used by the proxy, because the Rust crate for the proxy API bindings
doesn't actually include the generated code for the new `outbound`
proto.

This branch adds a module including the generated outbound API in the
Rust crate so that the new API can be used.