dst: add `Opaque` hint to `destination.ProtocolHint ` (#197)
This indicates that the destination will handle the connection as an
opaque TCP byte stream, and that HTTP/2 upgrades should not be
performed. If the `opaque_transport` flag is also set for this
destination, then the proxy should not send a session protocol in its
transport header.

See linkerd/linkerd2#9888 for details.