Add opaque route skeleton to #211 (#212)
The new API needs to be able to describe, generally, opaque routes, even
though there is no user-facing configuration for these routes at the
moment.

This change eliminates the `Backend` on the top level response--it was
used only to synthesize an opaque route in the proxy. This opaque
route should be represented properly in the API and synthesized in the
control plane.