feat(destination): add resource metadata to responses
The destination service does no include resource coordinates, so the proxy must
"guess" them by parsing hostnames, which is not particularly reliable or
efficient.

This change updates the destination API so that profile responses may include
resource reference to (1) parent Services, (2) ServiceProfile resources and (3)
backend Services. Furthermore, the WeightedAddr type is updated to include a
resource reference so that Pod coordinates may be referenced.

This metadata is not required to be set (for backwards compatibility).